Moved all dual quaternion functionality from GTC space to GTX.
