+++ test/gtc/CMakeLists.txt
@@ -7,7 +7,6 @@ glmCreateTestGTC(gtc_matrix_inverse)
 glmCreateTestGTC(gtc_matrix_transform)
 glmCreateTestGTC(gtc_noise)
 glmCreateTestGTC(gtc_quaternion)
-glmCreateTestGTC(gtc_dual_quaternion)
 glmCreateTestGTC(gtc_random)
 glmCreateTestGTC(gtc_reciprocal)
 glmCreateTestGTC(gtc_swizzle)
